Respects robots.txt (in a simplified way), filters what links to follow cleaned the code, fixed small details

+++ localindices-harvester/src/java/com/indexdata/masterkey/localindices/harvest/job/WebHarvestJob.java
@@ -5,9 +5,12 @@ package com.indexdata.masterkey.localindices.harvest.job;
 
 import com.indexdata.masterkey.localindices.entity.WebCrawlResource;
 import com.indexdata.masterkey.localindices.harvest.storage.HarvestStorage;
+import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.*;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.Vector;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -20,12 +23,22 @@ import org.apache.log4j.Logger;
  * @author heikki
  * 
  * TODO:
- *   - filter what types of links to follow
- *   - Robots.txt (with cache)
- *   - Redirects (watch out for loops etc)
- *   - better load reduction for servers.
+ *   - filter what links to follow
+ *   - Redirects (watch out for loops etc) (doesn't the library handle this?)
  *   - Get text into zebra
- *   - error handling, time outs, etc
+ *   - error handling, 
+ *   - Parameters
+ *     - Remove the filetype mask
+ *     - Add max pages to harvest
+ * TODO - tests
+ *   - Test a site with plain text files
+ *   - Test with a tarpit-like site
+ * TODO - but in some later version
+ *   - robots.txt
+ *     - check user-agent lines
+ *     - understand also Allow lines
+ *   - better load reduction for servers.
+ *     - keep todo list in a priority queue, with time stamps when a hosts turn is
  */
 public class WebHarvestJob implements HarvestJob {
 
@@ -35,19 +48,27 @@ public class WebHarvestJob implements HarvestJob {
     private String error;
     private WebCrawlResource resource;
     private boolean die = false;
-    private Vector<String> toSearch; // todo list for this round
-    private Vector<String> searched; // all pages we have seen
-    private Vector<String> nextRound; // links found in this round are pushed here
+    private Vector<URL> toSearch; // todo list for this round
+    private Vector<URL> searched; // all pages we have seen
+    private Vector<URL> nextRound; // links found in this round are pushed here
     private int numtosearch; 
     private int round;
+    private Map<URL,String> robotCache = new HashMap<URL,String>();
+    
+    private final static int connTimeOut = 3000; // ms to make a connection
+    private final static int readTimeOut = 3000; // ms to read a block
+      // About 30 seconds seems reasonable. 
+    private final static int readBlockSize = 1000000; // bytes to read in one op
+    private final static int maxReadSize =  10000000; // 10MB 
+    private final static String userAgentString = "IndexData Masterkey Web crawler";
 
     private class pageInfo {
 
-        public String url;
+        public URL url;
         public String content; // the whole thing
         public String headers;
         public String body;
-        public Vector<String> links;
+        public Vector<URL> links;
         public String plaintext;
         public String title;
     }
@@ -56,6 +77,7 @@ public class WebHarvestJob implements HarvestJob {
         this.resource = resource;
         this.status = HarvestStatus.NEW;
         logger.setLevel(Level.ALL);  // While debugging
+        this.error=null;    
     }
 
     private synchronized boolean isKillSendt() {
@@ -103,24 +125,20 @@ public class WebHarvestJob implements HarvestJob {
         return error;
     }
 
-    private String fetchPage(String pageUrl) {
+    private String fetchPage(URL url) {
         // FIXME - Return a pageinfo structure with all the data and error messages
-        // FIXME - Take the URL as an URL, not as a string
+        String pageUrl = url.toString();
         logger.log(Level.TRACE, "About to fetch '" + pageUrl + "'");
-        URL url;
-        try {
-            url = new URL(pageUrl);
-        } catch (MalformedURLException e) {
-            logger.log(Level.ERROR, "Invalid URL: '" + pageUrl + "': " + e.getMessage());
-            return "";
-        }
         if (url.getProtocol().compareTo("http") != 0) {
             logger.log(Level.ERROR, "Unsupported protocol in '" + pageUrl);
             return "";
         }
         try {
             URLConnection urlConnection = url.openConnection();
+            urlConnection.setConnectTimeout(connTimeOut);
+            urlConnection.setReadTimeout(readTimeOut);
             urlConnection.setAllowUserInteraction(false);
+            urlConnection.setRequestProperty("User-agent",userAgentString );
             InputStream urlStream = url.openStream();
             String type = "";
             type = urlConnection.getContentType();
@@ -131,19 +149,22 @@ public class WebHarvestJob implements HarvestJob {
                 logger.log(Level.DEBUG, pageUrl + " skipped: could not guess type");
                 return "";
             }
-            if (!type.startsWith("text/html")) {
+            if (!type.startsWith("text/html") &&
+                !type.startsWith("text/plain")  ) {
+                // Get also plain text, we need it for robots.txt, and
+                // might as well index it all anyway
                 logger.log(Level.DEBUG, pageUrl + " skipped: type '" +
-                        type + "' not text/html");
+                        type + "' not text/html (nor text/plain)");
                 return "";
             }
             // search the input stream for links
             // first, read in the entire URL
-            byte b[] = new byte[1000];
+            byte b[] = new byte[readBlockSize];
             int numRead = urlStream.read(b);
             String content = new String(b, 0, numRead);
             //FIXME - Check if content all too big, abort! Could be a tar pit
             // or misbehaving web server that just keeps outputting rubbish
-            while (numRead != -1) {
+            while ((numRead != -1) && ( content.length() < maxReadSize) ) {
                 numRead = urlStream.read(b);
                 if (numRead != -1) {
                     String newContent = new String(b, 0, numRead);
@@ -154,18 +175,21 @@ public class WebHarvestJob implements HarvestJob {
             logger.log(Level.TRACE, pageUrl + " Read " + content.length() + " bytes");
             //logger.log(Level.DEBUG, content );
             return content;
+        } catch ( FileNotFoundException ex ) {
+            logger.log(Level.ERROR, "I/O Exception: Not found " + pageUrl );
         } catch (IOException ex) {
-            logger.log(Level.ERROR, "I/O Exception with " + pageUrl + ": " +
-                    ex.getMessage());
-            return "";
+            logger.log(Level.ERROR, "I/O Exception " +
+                    "(" + ex.getClass().getSimpleName() +") "+
+                    "with " + pageUrl + ": " + ex.getMessage());
         }
+        return ""; // signals an error, real return is inside the try block
     } // fetchPage
 
-    private pageInfo splitPage(String page, String url) {
+    private pageInfo splitPage(String page, URL pageUrl) {
         pageInfo pi = new pageInfo();
-        pi.links = new Vector<String>();
+        pi.links = new Vector<URL>();
         pi.content = page; // save it all for future reference
-        pi.url = url;
+        pi.url = pageUrl;
         // Split headers and body, if possible
         Pattern p1 = Pattern.compile("<head>(.*)</head>.*" +
                 "<body[^>]*>(.*)",
@@ -194,12 +218,6 @@ public class WebHarvestJob implements HarvestJob {
         // extract full text
 
         // extract links
-        URL pageUrl;
-        try {
-            pageUrl = new URL(url);
-        } catch (MalformedURLException ex) {
-            return pi; // without extracting links, they won't be good without this
-        }
 
         //Pattern p3 = Pattern.compile("<a([^>]+)>",
         Pattern p3 = Pattern.compile("<a[^>]+href=['\"]?([^>'\"#]+)['\"# ]?[^>]*>",
@@ -215,34 +233,104 @@ public class WebHarvestJob implements HarvestJob {
                 logger.log(Level.TRACE, "Could not make a good url from '" + lnk + "'");
                 break;
             }
-            String newlnk = linkUrl.toString();
             //logger.log(Level.TRACE, "Found link '" + m3.group() + "' '" + lnk + "' " +
-            //        "-> '" + newlnk + "'");
-            if (!pi.links.contains(newlnk)) {
-                pi.links.add(newlnk);
+            //        "-> '" + linkUrl.toString() + "'");
+            if (!pi.links.contains(linkUrl)) {
+                pi.links.add(linkUrl);
             }
         }
 
         logger.log(Level.DEBUG, 
-                resource.getName() +
-                " " + round +":" +searched.size() +"/" + numtosearch +
-                " " + pi.url + " " +
+                "JOB#" + resource.getId() + " " + 
+                round +":" +searched.size() +"/" + numtosearch + " " + 
+                pi.url + " " +
                 "title:'" + pi.title + "' (" +
                 "h=" + pi.headers.length() + "b " +
                 "b=" + pi.body.length() + "b) " +
                 pi.links.size() + " links");
-
-
         return pi;
     }
 
+    private boolean checkRobots(URL url) {
+        String strHost = url.getHost();
+        if (strHost==null || strHost.isEmpty() ) {
+            // Should not happen!
+            logger.log(Level.DEBUG, "Could not extract host from '" +
+                    url.toString()+"' - skipping robots txt");
+	    return false;            
+        }
+	String strRobot = "http://" + strHost + "/robots.txt";
+	URL robUrl;
+	try { 
+	    robUrl = new URL(strRobot);
+	} catch (MalformedURLException e) {
+	    // something weird is happening, so don't trust it
+            logger.log(Level.DEBUG, "Could not create robot url "+
+                    "'" + strRobot +"'");            
+	    return false;
+	}
+        String robtxt=robotCache.get(robUrl);
+        if ( robtxt == null ) {
+            robtxt = fetchPage(robUrl);
+            if (robtxt==null)
+                robtxt="";
+            robotCache.put(robUrl, robtxt); 
+            logger.log(Level.DEBUG, "Got " + robUrl.toString()+ "\n" +
+                    robtxt);
+        }
+        if (robtxt.isEmpty())
+            return true; // no robots.txt, go ahead
+        
+        // Simplified, we assume all User-agent lines apply to us
+        // Most likely nobody has (yet?) written a robots.txt section 
+        // for specifically for us.
+        Pattern p = Pattern.compile("^Disallow:\\s*(.*\\S)\\s*$",
+                Pattern.CASE_INSENSITIVE | Pattern.MULTILINE );
+        Matcher m = p.matcher(robtxt);
+        String urlpath=url.getPath();
+        while (m.find()) {
+            String path = m.group(1);
+            if (urlpath.startsWith(path) ) {
+                logger.log(Level.TRACE, "Path '"+urlpath+"' forbidden "+
+                        "by robot '" + path +"'");                
+                return false; // found one they don't want us to go to
+            }
+        }
+        //logger.log(Level.TRACE, "Path '"+urlpath+"' all right ");
+        return true;
+    }
+    
+    private boolean filterLink(URL url) {
+        if ( resource.getUriMasks().isEmpty() ) {
+            return true; // no filtering, go anywhere
+        }
+        String urlStr = url.toString();
+        for (String mask : resource.getUriMasks().split(" ") ) {
+            Pattern p = Pattern.compile(mask);
+            Matcher m = p.matcher(urlStr);
+            if ( m.find() ) {
+                logger.log(Level.TRACE, "url '"+urlStr+"' matched '" + mask + "'");
+                return true;
+            }
+        }
+        logger.log(Level.TRACE, "url '"+urlStr+"' refused by filters");
+        return false;
+    }
+    
     private void initWorkList() {
-        searched = new Vector<String>();
-        toSearch = new Vector<String>();
-        nextRound = new Vector<String>();
+        searched = new Vector<URL>();
+        toSearch = new Vector<URL>();
+        nextRound = new Vector<URL>();
         for (String startUrl : resource.getStartUrls().split(" ")) {
-            toSearch.add(startUrl);
-            logger.log(Level.TRACE, "Added start URL '" + startUrl + "'");
+            URL url;
+            try {
+                url = new URL(startUrl);
+                toSearch.add(url);
+                logger.log(Level.TRACE, "Added start URL '" + startUrl + "'");
+            } catch (MalformedURLException ex) {
+                setError("Invalid start url: '" + startUrl+"'" );
+                toSearch.clear();
+            }
         // Fixme - validate they all are OK
         }
         numtosearch = toSearch.size();
@@ -251,17 +339,18 @@ public class WebHarvestJob implements HarvestJob {
     /** Harvest one round of links */
     private void harvestRound() {
         while (!toSearch.isEmpty() && !isKillSendt()) {
-            String curUrl = toSearch.firstElement();
+            URL curUrl = toSearch.firstElement();
             toSearch.removeElementAt(0);
             if (!searched.contains(curUrl)) {
                 searched.add(curUrl); // to make sure we don't go there again  
-                // Fixme - check robot safe
-                resource.setCurrentStatus("" + searched.size() + "pages visited");
+                if ( !checkRobots(curUrl))
+                    break;
+                //resource.setCurrentStatus("" + searched.size() + " pages visited");
                 String page = fetchPage(curUrl);
                 if (!page.isEmpty()) {
                     pageInfo pi = splitPage(page, curUrl);
-                    for (String u : pi.links) {
-                        if (!nextRound.contains(u)) {
+                    for (URL u : pi.links) {
+                        if (!nextRound.contains(u) && filterLink(u)) {
                             // FIXME - filter those links that need to be followed
                             nextRound.add(u);
                         }
@@ -280,8 +369,11 @@ public class WebHarvestJob implements HarvestJob {
     /** Harvest all there is to do */
     private void harvestLoop() {
         round = 0;
+        long startTime = System.currentTimeMillis();
         initWorkList();
-        while (round <= resource.getRecursionDepth() && !isKillSendt()) {
+        while (round <= resource.getRecursionDepth() 
+                && !toSearch.isEmpty() 
+                && !isKillSendt()) {
             harvestRound();
             logger.log(Level.DEBUG, "Round " + round + ": " +
                     searched.size() + " urls seen. " +
@@ -291,20 +383,27 @@ public class WebHarvestJob implements HarvestJob {
             nextRound.clear();
             round++;
         }
-        logger.log(Level.DEBUG, "Did " + (round - 1) + " rounds. " +
-                "(max=" + resource.getRecursionDepth() + ") " +
-                "(iskilled=" + isKillSendt() + ") " +
-                "Seen " + searched.size() + " urls");
+        long elapsed = (System.currentTimeMillis() - startTime) / 1000; // sec
+        
+        String killmsg="Did";
+        if (isKillSendt()) 
+            killmsg="Killed after";
+        logger.log(Level.DEBUG, killmsg +" " + (round - 1) + " rounds. " +
+                "Seen " + searched.size() + " urls " +
+                " in " + elapsed + " seconds " +
+                "(Next depth would have taken " + toSearch.size() + " pages more)");
     }
 
     public void run() {
         status = HarvestStatus.RUNNING;
         harvestLoop();
-        if (isKillSendt()) {
-            setError("Web Crawl interrupted with a kill signal");
-        } else {
-            //status = HarvestStatus.FINISHED;
-            setError("All done - but we call it an error so we can do again");
+        if (this.error == null) {
+            if (isKillSendt()) {
+                setError("Web Crawl interrupted with a kill signal");
+            } else {
+                //status = HarvestStatus.FINISHED;
+                setError("All done - but we call it an error so we can do again");
+            }
         }
     } // run()
 }
